SLIPS V2/requirements.txt
# Core dependencies
redis>=4.3.4
pyyaml>=6.0
requests>=2.28.1
ipaddress>=1.0.23

# ML dependencies (optional)
tensorflow>=2.10.0; platform_machine != "armv7l"
scikit-learn>=1.1.2
joblib>=1.1.0
numpy>=1.23.2

# Utilities
colorama>=0.4.5
watchdog>=2.1.9
python-dateutil>=2.8.2
